[REF] Separete MDe class
