+++ src/erpbrasil/edoc/edoc.py
@@ -75,15 +75,8 @@ class DocumentoEletronico(ABC):
 
         xml_string, xml_etree = self._generateds_to_string_etree(raiz)
         with self._transmissao.cliente(url):
-            # Recupera a sigla do estado
-            uf_list = [uf for nUF, uf in SIGLA_ESTADO.items() if
-                       nUF == str(getattr(raiz, 'cUFAutor', ''))]
-            if uf_list:
-                kwargs = dict(uf=uf_list and uf_list[0])
-            else:
-                kwargs = {}
             retorno = self._transmissao.enviar(
-                operacao, xml_etree, **kwargs
+                operacao, xml_etree
             )
             return analisar_retorno_raw(
                 operacao, raiz, xml_string, retorno, classe
