+++ src/erpbrasil/edoc/nfe.py
@@ -22,15 +22,6 @@ try:
     from nfelib.v4_00 import retEnvEvento
     from nfelib.v4_00 import retEnvEventoCancNFe
     from nfelib.v4_00 import retEnviNFe
-
-
-    from nfelib.v4_00 import leiauteConfRecebto
-    from nfelib.v4_00.leiauteConfRecebtoManifestacao import tpEventoType as eventoManifestacao  # noga
-    from nfelib.v4_00.leiauteConfRecebtoManifestacao import descEventoType as descEventoManifestacao  # noga
-    from nfelib.v4_00.leiauteConfRecebtoManifestacao import infEventoType as infEventoManifestacao  # noga
-    from nfelib.v4_00.leiauteConfRecebtoManifestacao import detEventoType as detEventoManifestacao  # noga
-    from nfelib.v4_00.leiauteConfRecebtoManifestacao import TEvento as TEventoManifestacao  # noga
-    from nfelib.v4_00.leiauteConfRecebtoManifestacao import TEnvEvento as TEnvEventoManifestacao  # noga
 except ImportError:
     pass
 
@@ -1021,160 +1012,4 @@ class NFe(DocumentoEletronico):
                           int(self.ambiente)),
             'nfeDistDFeInteresse',
             retDistDFeInt
-        )
-
-    # ----------------------------- MANIFESTAÇÃO DO DESTINATÁRIO -----------------
-
-    def nfe_recepcao_envia_lote_evento(self, lista_eventos, numero_lote=False):
-        """
-        Envia lote de eventos confRecebto.TEvento
-        :param lista_eventos: Lista com os eventos (infEvento)
-        :param numero_lote: Número do lote. Gerado caso None
-        :return: retorna a resposta do _post() de envio do lote
-        """
-
-        if not numero_lote:
-            numero_lote = self._gera_numero_lote()
-
-        eventos = []
-        raiz = TEnvEventoManifestacao(
-            versao="1.00",
-            idLote=numero_lote,
-            evento=eventos
-        )
-        raiz.original_tagname_ = 'envEvento'
-
-        for raiz_evento in lista_eventos:
-            evento = TEventoManifestacao(
-                versao="1.00", infEvento=raiz_evento,
-            )
-            evento.original_tagname_ = 'evento'
-
-            # Recupera o evento do XML assinado
-            xml_assinado = self.assina_raiz(evento, evento.infEvento.Id
-                                            ).replace('\n', '').encode()
-
-            # Converte o xml_assinado para um objeto pelo
-            # parser do esquema leiauteConfRecebto
-            xml_object = leiauteConfRecebto.parseString(xml_assinado)
-
-            # Adiciona o xml_object na lista de eventos. Desse modo a lista
-            # de eventos terá um evento assinado corretamente
-            eventos.append(xml_object)
-
-        return self._post(
-            raiz,
-            # 'https://homologacao.nfe.fazenda.sp.gov.br/ws/nferecepcaoevento4.asmx?wsdl',
-            localizar_url(WS_NFE_RECEPCAO_EVENTO, str(91), self.mod,
-                          int(self.ambiente)),
-            'nfeRecepcaoEventoNF',
-            retEnvEvento
-        )
-
-    def nfe_recepcao_monta_evento(self, chave, cnpj_cpf, tpEvento, descEvento,
-                                  dhEvento=None, xJust=None):
-        """
-        Método para montar o evento(infEvento) da manifestação
-        :param chave: chave do documento
-        :param cnpj_cpf: CPF ou CNPJ
-        :param tpEvento:   Código do Evento
-                                210200 – Confirmação da Operação
-                                210210 – Ciência da Operação
-                                210220 – Desconhecimento da Operação
-                                210240 – Operação não Realizada
-        :param descEvento: Informar a descrição do evento:
-                                Confirmacao da Operacao
-                                Ciencia da Operacao
-                                Desconhecimento da Operacao
-                                Operacao nao Realizada
-        :param dhEvento:   Data/Hora no formato AAAA-MM-DDThh:mm:ssTZD
-        :param xJust:      Este campo deve ser informado somente no
-                                evento de Operação não realizada
-        :return: Um objeto da classe confRecebto.infEventoType preenchido
-        """
-
-        nSeqEvento = '1'
-        raiz = infEventoManifestacao(
-            Id='ID{}{}{}'.format(tpEvento, chave, nSeqEvento.zfill(2)),
-            cOrgao=91,
-            tpAmb=self.ambiente,
-            CNPJ=cnpj_cpf if len(cnpj_cpf) > 11 else None,
-            CPF=cnpj_cpf if len(cnpj_cpf) <= 11 else None,
-            chNFe=chave,
-            dhEvento=dhEvento or self._hora_agora(),
-            tpEvento=tpEvento,
-            nSeqEvento=nSeqEvento,
-            verEvento='1.00',
-            detEvento=detEventoManifestacao(
-                versao='1.00',
-                descEvento=descEvento,
-                xJust=xJust
-            )
-        )
-
-        raiz.original_tagname_ = 'infEvento'
-
-        return raiz
-
-    def nfe_recepcao_evento(self, chave, cnpj_cpf, tpEvento, descEvento, xJust=None):
-        """
-        Envia a manifestação do destinatário para o WS
-        :param cnpj_cpf:   CPF ou CNPJ
-        :param tpEvento:   Código do Evento
-                             210200 – Confirmação da Operação
-                             210210 – Ciência da Operação
-                             210220 – Desconhecimento da Operação
-                             210240 – Operação não Realizada
-        :param descEvento: Informar a descrição do evento:
-                             Confirmacao da Operacao
-                             Ciencia da Operacao
-                             Desconhecimento da Operacao
-                             Operacao nao Realizada
-        :param xJust:      Este campo deve ser informado somente no
-                             evento de Operação não realizada
-        :return:
-        """
-
-        evento = self.nfe_recepcao_monta_evento(
-            chave, cnpj_cpf, tpEvento, descEvento, xJust=xJust)
-
-        # TODO: Verificar possibilidade de adaptar e utilizar código existente
-        #  em self.enviar_lote_evento(lista_eventos=[evento]).
-        #  A única diferença é a classe utilizada pelo evento
-
-        return self.nfe_recepcao_envia_lote_evento(
-            lista_eventos=[evento], numero_lote='1'
-        )
-
-    def confirmacao_da_operacao(self, chave, cnpj_cpf):
-        return self.nfe_recepcao_evento(
-            chave,
-            cnpj_cpf,
-            eventoManifestacao._2_10200,
-            descEventoManifestacao.CONFIRMACAODA_OPERACAO,
-        )
-
-    def ciencia_da_operacao(self, chave, cnpj_cpf):
-        return self.nfe_recepcao_evento(
-            chave,
-            cnpj_cpf,
-            eventoManifestacao._2_10210,
-            descEventoManifestacao.CIENCIADA_OPERACAO,
-        )
-
-    def desconhecimento_da_operacao(self, chave, cnpj_cpf):
-        return self.nfe_recepcao_evento(
-            chave,
-            cnpj_cpf,
-            eventoManifestacao._2_10220,
-            descEventoManifestacao.DESCONHECIMENTODA_OPERACAO,
-        )
-
-    def operacao_nao_realizada(self, chave, cnpj_cpf):
-        return self.nfe_recepcao_evento(
-            chave,
-            cnpj_cpf,
-            eventoManifestacao._2_10240,
-            descEventoManifestacao.OPERACAONAO_REALIZADA,
-            xJust=''.zfill(15)
         )
